[INTERPRETER] Fix integer overflow in tl.sum (#11531)
Under `TRITON_INTERPRET=1`, a scalar integer reduction can raise
`OverflowError` when its result exceeds the target dtype, while the
compiled kernel wraps around.

```python
@triton.jit
def kernel(X, Z, SIZE: tl.constexpr):
    x = tl.load(X + tl.arange(0, SIZE))
    tl.store(Z, tl.sum(x))

x = torch.full((128,), 2**30, dtype=torch.int32, device="cuda")
z = torch.zeros(1, dtype=torch.int32, device="cuda")
kernel[(1, )](x, z, SIZE=128)
```

produces:

```text
compiled:            0
TRITON_INTERPRET=1:  InterpreterError("OverflowError('Python integer 137438953472 out of bounds for int32')")
```

NumPy promotes the int32 accumulator to int64. In
`ReduceScanOpInterface.to_tensor`, shaped results are narrowed
separately with `astype`, which performs the expected fixed-width
integer conversion. The scalar path instead passed the target dtype
directly to `np.array`, which NumPy 2 rejects for an out-of-range
scalar.

Infer the intermediate dtype first and then narrow it with `astype`,
matching the shaped-result path. The NumPy deprecation warning that
preceded this NumPy 2 behavior recommended the same two-step conversion:
`np.array(value).astype(dtype)`.

#7470 worked around the same failure by limiting the integer inputs in
the reduce tests. Remove that restriction and compare integer sums
exactly, since a relative tolerance is inappropriate for wrapped integer
results. Reverting only the interpreter change makes the
`test_reduce1d[sum-int32-*]` cases fail again.

+++ python/triton/runtime/interpreter.py
@@ -1039,7 +1039,9 @@ class ReduceScanOpInterface:
             ret = ret.astype(np_dtype)
             ret_type = tl.block_type(dtype, list(ret.shape))
         else:
-            ret = np.array([ret], dtype=np_dtype)
+            # Match the shaped path above. NumPy 2 rejects out-of-range Python
+            # integers in array(..., dtype=...), while astype narrows instead.
+            ret = np.array([ret]).astype(np_dtype)
             ret_type = dtype
         return tl.core.tensor(TensorHandle(ret, dtype.scalar), ret_type)
